tests/wpt/web-platform-tests/css/css-box/parsing/width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing width with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-width">
<meta name="assert" content="width supports the full grammar '[<length> | <percentage>] | available | min-content | max-content | fit-content | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("width", "auto");

test_valid_value("width", "10px");
test_valid_value("width", "20%");
test_valid_value("width", "calc(2em + 3ex)");

test_valid_value("width", "min-content");
test_valid_value("width", "max-content");
test_valid_value("width", "fit-content");

// The following is not yet supported by browsers:
// test_valid_value("width", "available");
</script>
</body>
</html>
